Moved FileStoreException class
