+++ ambry-store/src/test/java/com/github/ambry/store/FileStoreTest.java
@@ -13,13 +13,11 @@
  */
 package com.github.ambry.store;
 
-import com.github.ambry.clustermap.FileStoreException;
 import com.github.ambry.config.FileCopyConfig;
 import com.github.ambry.config.VerifiableProperties;
 import java.io.DataInputStream;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.RandomAccessFile;
